refactor(sim): split delegation spoofing facade
